% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jobcontrolFunctions.R
\name{delete_job}
\alias{delete_job}
\title{Delete all the chunks associated with a specific job}
\usage{
delete_job(job_id, con = NULL, ...)
}
\arguments{
\item{job_id}{The id of the job from which to delete all chunks}

\item{con}{The database connection to use for the deletion}

\item{...}{Other arguments to pass to the connection}
}
\value{
TRUE on success
}
\description{
Delete all the chunks associated with a specific job
}

