% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jobcontrolFunctions.R
\name{get_chunk}
\alias{get_chunk}
\title{Get a chunk of work to do from a specified job}
\usage{
get_chunk(worker_id, job_id = 1, worker_name = paste(Sys.info()["nodename"],
  worker_id, sep = ":"), nchunks = 1, con = NULL,
  nullchunk = NA_character_, ...)
}
\arguments{
\item{worker_id}{The id of the worker requesting a chunk}

\item{job_id}{The id of the job for which chunks are being requested}

\item{worker_name}{The name of the worker requesting a chunk. Default:
\code{<nodename>:<worker_id>}}

\item{nchunks}{The number of chunks requested}

\item{con}{The database connection to use for the chunk request or
\code{NULL} to indicate that the function should make and close a
connection just for this request.}

\item{nullchunk}{The object to return if no chunks are returned}

\item{...}{Other arguments for creating the connection when \code{NULL}}
}
\value{
A data.frame with \code{nchunks} rows (some of which may contain
  \code{NA}s) or if no chunks at all were available a matrix with
  \code{nchunks} rows and 1 column, filled with \code{NA}s.
}
\description{
Get a chunk of work to do from a specified job
}

