% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jobcontrolFunctions.R
\name{.jobcontrol_connection}
\alias{.jobcontrol_connection}
\title{Create a connection to the job control database}
\usage{
.jobcontrol_connection(con = NULL, group = "Rgridengineswarm", ...)
}
\arguments{
\item{group}{The group from default.file to use for the database connection}

\item{...}{Other arguments to pass to the connection}
}
\description{
Create a connection to the job control database
}

