% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Rgridengineswarm-package.r
\docType{package}
\name{Rgridengineswarm}
\alias{Rgridengineswarm}
\alias{Rgridengineswarm-package}
\title{Rgridengineswarm}
\description{
Rgridengineswarm
}
\section{Package Options}{


You can change the details of how Rgridengineswarm connects to the job control
database by setting

\verb{options(Rgridengineswarm.connpararams=list(<db connection params>))}

see \code{\link{.jobcontrol_connection}} for details.
}

